% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baffle.r
\docType{package}
\name{baffle-package}
\alias{baffle}
\alias{baffle-package}
\title{Baffle: Waffle graphs with base graphics}
\description{
The package provides functions to make waffle graphs in the base graphics.
}
\details{
Waffle plots are rectangular pie charts that represent a quantity or abundances using
colored squares or other symbol. This makes them better at transmitting information as the
discrete number of squares is easier to read than the circular area of pie charts.
While the original waffle charts were rectangular with 10 rows and columns, with a single square
representing 1%, they are nowadays popular in various infographics to visualize any
proportional ratios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://j-moravec.github.io/baffle/}
  \item \url{https://github.com/j-moravec/baffle}
  \item Report bugs at \url{https://github.com/j-moravec/baffle/issues}
}

}
\author{
\strong{Maintainer}: Jiří Moravec \email{jiri.c.moravec@gmail.com}

}
\keyword{internal}
