% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baffle.R
\name{baffle-shiny}
\alias{baffle-shiny}
\alias{baffleOutput}
\alias{renderBaffle}
\alias{baffle_proxy}
\title{Shiny bindings for baffle}
\usage{
baffleOutput(outputId, width = "100\%", height = "400px")

renderBaffle(expr, env = parent.frame(), quoted = FALSE)

baffle_proxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a baffle.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id, outputId}{Output variable to read from.}

\item{session}{Shiny session.}
}
\description{
Output and render functions for using baffle within Shiny
applications and interactive Rmd documents.
}
