% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{proxy}
\alias{proxy}
\alias{b_once_p}
\alias{b_start_p}
\alias{b_stop_p}
\alias{b_reveal_p}
\title{Proxies}
\usage{
b_once_p(proxy, delay = 0)

b_start_p(proxy, delay = 0)

b_stop_p(proxy, delay = 0)

b_reveal_p(proxy, delay = NULL)
}
\arguments{
\item{proxy}{A shiny proxy as returned by \code{\link{baffle_proxy}}.}

\item{delay}{Delay in milliseconds before action is triggered.}
}
\description{
Shiny proxies.
}
\examples{
library(shiny)

ui <- fluidPage(
  baffleOutput("baffle"),
  br(),
  actionButton("start", "baffle")
)

server <- function(input, output, session){

  output$baffle <- renderBaffle({
    b_affle("baffle")
  })

  observeEvent(input$start, {
    baffle_proxy("baffle") \%>\%
      b_start_p() \%>\%
      b_reveal_p(3000)
  })

}

if(interactive())
  shinyApp(ui, server)

}
