% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle.r
\name{waffle}
\alias{waffle}
\alias{waffle.mat}
\title{Make a waffle chart.}
\usage{
waffle(
  x,
  col,
  nrow = 10,
  ncol = 10,
  border = NA,
  adj = 0.1,
  byrow = TRUE,
  add = FALSE
)

waffle.mat(x, col, border = NA, adj = 0.1, add = FALSE)
}
\arguments{
\item{x}{a vector of abundances}

\item{col}{a vector of colors, must be the same length as the x. If not specified, the "Set1"
palette from the package RColorBrewer is used.}

\item{nrow}{an amount of rows}

\item{ncol}{an amount of columns}

\item{border}{the border color of squares, by default no border is plotted}

\item{adj}{an amount by which squares are shrinked so they do not touch each other}

\item{byrow}{whether to fill matrix by rows (default) or columns}

\item{add}{whether to add to a current plot}

\item{mat}{a design matrix (see details)}
}
\description{
Waffle chart is a pie chart that visually represents abundances by the number of squares.
}
\details{
The waffle chart consists of squares plotted on a rectangular lattice according to a design
matrix that is constructed automatically from a vector of abundances or can be directly provided.

The \strong{waffle} function accepts a vector of abundances and constructs the design matrix,
which is then parsed to the \code{waffle.mat} function. By default, the design matrix is by row
with sequential numbers according to the vector of abundances \code{x}. This means that if
the \code{x=c(3,5)}, the matrix will be filled with three \code{1} and five \code{2}, all other
cells of the matrix are set as unknown values \code{NA}. By default, 10 rows and 10 columns are
used so that one square represents 1% of the total. This might have an undesirable side-effect
If the sum of values in \code{x} is significantly smaller than 100 as a relatively large amount
of empty space will still be allocated. This can be easily corrected by providing a reasonable
number of rows and columns.

The function \strong{waffle.mat} accepts a custom-made design matrix and thus allows a better
control of colored regions. It is called internally by the \code{waffle} function, which serves
as an easy to use interface for the \code{waffle.mat}. For this reason, the \code{waffle.mat}
does not checks for the validity of input arguments and does not set a default colors.

The assumed and allocated coordinate system is from \code{1} to \code{ncol+1} for x and
from \code{1} to \code{ncol+1} for y. Squares are filled from top right corner of this coordinate
system.

If \code{add=FALSE}, a new window with a fixed aspect ratio x/y=1 is allocated so that
plotted rectangles are squares. This might cause the plot margins, and thus the main title,
to be quite far away. In this case, plotting the title using \code{text} instead of \code{title}
might be a better idea. In this case, the coordinates might be:
\code{text(x=(ncol+2)/2, y=nrow+1,...)}
}
\examples{
waffle(c(50,25,25))
waffle(c(25,75), c("darkorchid", "lightgray"))
waffle(c(14,8,4), nrow=3)

# custom design matrix with a more complex structure
cols = RColorBrewer::brewer.pal(3, "Set1")
design_mat = matrix(NA, 7,10)
design_mat[1,] = 1
design_mat[2,1:8] = 1
design_mat[4,] = 2
design_mat[5,1:4] = 2
design_mat[7,1:5] = 3
waffle.mat(design_mat, cols)

}
\seealso{
\code{\link[waffle]{waffle}} for a ggplot2 version.
}
